requirements.txt
ag2==0.9.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.7
autogen-agentchat==0.7.1
autogen-core==0.7.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
diskcache==5.6.3
docker==7.1.0
faiss-cpu==1.11.0.post1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google_search_results==2.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-ollama==0.3.6
langchain-text-splitters==0.3.9
langchainhub==0.1.21
langgraph==0.6.0
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.0
langgraph-sdk==0.2.0
langsmith==0.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==2.0.1
numpy==2.3.2
ollama==0.5.1
opentelemetry-api==1.35.0
orjson==3.11.1
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
pyarrow==21.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pypdf==5.9.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
serpapi==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.47.1
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
